+++ Home.py
@@ -6,6 +6,7 @@ import os
 from logzero import logger
 
 import helper.database_functions as db_funcs
+import helper.utils as utils
 
 os.environ["OAUTHLIB_INSECURE_TRANSPORT"] = "1"
 os.environ['GOOGLE_CLIENT_ID'] = st.secrets['google_oauth']['client_id']
@@ -22,7 +23,7 @@ flow = Flow.from_client_config(
             "redirect_uris": st.secrets["google_oauth"]["redirect_uris"]
         }
     },
-    scopes=['openid', 'https://www.googleapis.com/auth/userinfo.email', 'https://www.googleapis.com/auth/userinfo.profile', 'https://www.googleapis.com/auth/calendar'],
+    scopes=['openid', 'https://www.googleapis.com/auth/userinfo.email', 'https://www.googleapis.com/auth/userinfo.profile', 'https://www.googleapis.com/auth/calendar','https://www.googleapis.com/auth/tasks'],
     redirect_uri= st.secrets["google_oauth"]["redirect_uris"][0]
 )
 
@@ -54,13 +55,13 @@ def initial_display_elements():
     st.markdown("""
         <style>
             h2 {
-                color: #FF0000; /* Blue color for h2 headings */
+                color: #1E90FF; /* Blue color for h2 headings */
             }
             h3 {
-                color: #FF0000; /* Blue color for h3 headings */
+                color: #1E90FF; /* Blue color for h3 headings */
             }
             h4 {
-                color: #FF0000; /* Blue color for h3 headings */
+                color: #1E90FF; /* Blue color for h3 headings */
             }
             p, li {
                 color: #FFFFFF; /* White color for paragraph and list items */
@@ -68,9 +69,8 @@ def initial_display_elements():
         </style>
         """, unsafe_allow_html=True)
     
-    st.header(":red[TaskBreakdown - Development] 📝", divider='rainbow')
-    # st.sidebar.markdown("# Home 📝")
-    st.subheader("Break down your big goals into manageable steps")
+    st.header("TaskBreakdown 📝", divider='rainbow')
+    st.write("Break down your big goals into manageable steps")
 
     st.markdown("""
     ## About TaskBreakdown
@@ -79,7 +79,8 @@ def initial_display_elements():
     #### What Can You Do with TaskBreakdown?
     - **Set Long-Term Goals**: Define your big goals and let the assistant help you break them down.
     - **Create Action Plans**: Get detailed, actionable steps to achieve your goals.
-    - **Sync with Google Calendar**: Schedule your tasks and stay on track by syncing them with your Google Calendar.
+    - **Sync with Google Calendar**: Schedule your plan by syncing them with your Google Calendar.
+    - **Sync Tasks with Google Tasks**: Seamlessly add or update tasks in your Google Tasks to keep track of tasks to complete on a day/day basis
     - **Track Progress**: Use the structured plan to make tangible progress on your goals.
 
     #### Example Goals:
@@ -96,26 +97,40 @@ def initial_display_elements():
     3. **Navigate to the Todolist Tab**: Go to the Todolist tab using the side bar and read the How to Use pop down to get help with using the assistant.
     """)
 
-@st.cache_data(show_spinner=False)
-def get_cached_api_key(email:str):
-    db, cursor = db_funcs.initialize_database()
-    return db_funcs.get_user_api_key(cursor, email)
-
 if __name__ == "__main__":
+    st.set_page_config(page_title='Taskbreakdown', page_icon='', initial_sidebar_state='expanded', layout='wide', menu_items={'Report a Bug':'https://forms.gle/C8Zv8hzvYhPPvDW16'})
+    login_status_container = st.container()
     initial_display_elements()
     db, cursor = db_funcs.initialize_database()
 
     if 'user_info' not in st.session_state:
         st.session_state['credentials'] = None
         st.session_state['user_info'] = None
-        st.warning(body="You're not logged in, please login to use the assistant")
+        st.session_state['variables_initialised'] = False
 
     if st.session_state['user_info']:
         if not db_funcs.is_user_present(cursor, st.session_state['user_info']['email']):
-            db_funcs.save_user(cursor, db, st.session_state['user_info']['email'], st.session_state['user_info']['name'], st.session_state['user_info']['picture'])
-    
-        st.info(f"Welcome {st.session_state['user_info']['name']} ")
-        st.success("Setup Ready! You can now head onto the Todolist tab, to talk to the assistant :)")
+            db_funcs.save_user(cursor, db, st.session_state['user_info']['email'], st.session_state['user_info'].get('name', 'User'), st.session_state['user_info'].get('picture', ''))
+
+        if not st.session_state['variables_initialised']:
+            utils.initialize_variables()
+            st.session_state['calendar_service'] = utils.get_calendar_service()
+            st.session_state['timezone'] = utils.get_user_timezone(st.session_state['calendar_service'])
+            st.session_state['task_ids_generated'] = db_funcs.check_if_google_tasks_are_created(cursor, st.session_state['user_info']['email'])
+            st.session_state['plan'] = db_funcs.fetch_plan_if_generated(cursor, st.session_state['user_info']['email'])
+            if st.session_state['plan']:
+                first_entry = st.session_state['plan'][0]
+                last_entry = st.session_state['plan'][-1]
+        
+                # Initialize the variables
+                st.session_state['start_date'] = first_entry['date']
+                st.session_state['end_date'] = last_entry['date']
+                st.session_state['start_time'] = first_entry['start_time']
+                st.session_state['end_time'] = first_entry['end_time']
+        
+        with login_status_container:
+            st.success(f"Welcome {st.session_state['user_info']['name']}. Setup is ready! You can now head onto the Todolist tab, to talk to the assistant :)")
+        st.toast("Setup Ready! You can now head onto the Todolist tab, to talk to the assistant :)")
         logger.info(f"Welcome {st.session_state['user_info']['name']} ")
     else:
         user_info = process_callback()
@@ -124,7 +139,9 @@ if __name__ == "__main__":
             st.info(f"You've logged in as {user_info['name']}")
             st.rerun()
         else:
-            google_oauth()
+            with login_status_container:
+                st.warning(body="You're not logged in, please login to use the assistant")
+                google_oauth()
 
     st.markdown("""
     ---